Fix critical/high financial security issues
CRITICAL:
- Server-side math: recalculate lineTotal/subTotal/tax/total from raw
  line items in createOrder; client-supplied totals are ignored entirely
- Overpayment guard: reject request if paidAmount > server-computed total
- Toggle consistency: paidAmount is now kept in sync when toggling
  payment status (paid → paidAmount=total, unpaid → paidAmount=0),
  eliminating the invalid dueAmount=0 / paidAmount=0 ghost state

HIGH:
- DECIMAL(15,2): all monetary columns changed from DOUBLE to DECIMAL in
  every model (orders, orderItems, payments, customers, suppliers,
  purchaseBills, purchaseItems, dailySummary, ledger, stock, etc.)
- Ledger errors are now transaction-blocking when Chart of Accounts is
  initialized (order/payment/toggle creation rolls back on ledger failure
  instead of silently swallowing the error)
- Balance clamping removed: Math.max(0, ...) stripped from all
  customer/supplier currentBalance updates so credits are not silently
  dropped
- Reference integrity: createPayment now validates that referenceId
  (order or purchaseBill) exists and is not soft-deleted before
  inserting the payment record
- Payment reversal ledger: deletePayment ledger call is now blocking
  (rolls back the deletion if reversal journal fails)

MEDIUM:
- getOrder DAO: added isDeleted:false filter so soft-deleted invoices
  cannot be fetched via the direct DAO path

+++ backend/src/models/dailySummary.js
@@ -15,7 +15,7 @@ module.exports = (sequelize, Sequelize) => {
             },
             // Opening balance for the day (cash in drawer at start)
             openingBalance: {
-                type: Sequelize.DOUBLE,
+                type: Sequelize.DECIMAL(15, 2),
                 defaultValue: 0
             },
             openingBalanceSetAt: {
@@ -27,7 +27,7 @@ module.exports = (sequelize, Sequelize) => {
                 allowNull: true
             },
             totalSales: {
-                type: Sequelize.DOUBLE,
+                type: Sequelize.DECIMAL(15, 2),
                 defaultValue: 0
             },
             totalOrders: {
@@ -35,15 +35,15 @@ module.exports = (sequelize, Sequelize) => {
                 defaultValue: 0
             },
             totalPurchases: {
-                type: Sequelize.DOUBLE,
+                type: Sequelize.DECIMAL(15, 2),
                 defaultValue: 0
             },
             totalPaymentsReceived: {
-                type: Sequelize.DOUBLE,
+                type: Sequelize.DECIMAL(15, 2),
                 defaultValue: 0
             },
             totalPaymentsMade: {
-                type: Sequelize.DOUBLE,
+                type: Sequelize.DECIMAL(15, 2),
                 defaultValue: 0
             },
             // Running invoice number for the day
@@ -71,7 +71,7 @@ module.exports = (sequelize, Sequelize) => {
             },
             // Closing balance (calculated: opening + sales - expenses)
             closingBalance: {
-                type: Sequelize.DOUBLE,
+                type: Sequelize.DECIMAL(15, 2),
                 allowNull: true
             },
             notes: {
