+++ backend/src/models/payment.js
@@ -30,7 +30,7 @@ module.exports = (sequelize, Sequelize) => {
                 allowNull: false
             },
             amount: {
-                type: Sequelize.DOUBLE,
+                type: Sequelize.DECIMAL(15, 2),
                 allowNull: false
             },
             referenceType: {
